Remove file partition shims
